+++ python/make_ansible_portable/__init__.py
@@ -1,2 +1,2 @@
 TOOL_NAME = "make_ansible_portable"
-TOOL_VERSION = "0.2.0"
+TOOL_VERSION = "0.3.0"